Update IAM permissions in README.md
The two permissions previously listed are not enough to scrape the metrics. While we can build a detailed list of all the needed permissions, I think specifying just the role could be enough. The permission list might change in the future, and it isn't easy to maintain. Reasonably, the role will be updated by Google if additional permissions would be required.